requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyroscope-io>=0.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
